requirements.txt
python-dotenv>=1.0.0
nest-asyncio>=1.5.7
emoji>=2.8.0
pinecone>=3.0.0
semantic-router>=0.0.38

torch>=2.0.0

transformers>=4.35.0
pdfminer.six>=20221105
langchain>=0.0.267
langchain-openai>=0.0.1
langchain-community>=0.0.1

python-telegram-bot>=20.4
fastapi>=0.103.1
uvicorn>=0.23.2
python-multipart>=0.0.6

speechrecognition>=3.10.0
numpy>=1.24.0
rich>=13.5.3
groq>=0.4.0
deep-translator>=1.11.4
pydantic>=2.4.2